requirements.txt
akad
requests
rsa
PyQRCode
hyper
httpx
httplib2
python-axolotl-curve25519
pycryptodome